+++ pylatex/base_classes/__init__.py
@@ -6,7 +6,7 @@ Baseclasses that can be used to create classes representing LaTeX objects.
 """
 
 from .latex_object import LatexObject
-from .containers import Container, Environment, PreambleCommand
+from .containers import Container, Environment, ContainerCommand
 from .command import CommandBase, Command, UnsafeCommand, Options, \
     SpecialOptions, Arguments
 from .float import Float